refactor: extract useDecode hook into separate file
