+++ app/GRAPHQL/helpers/padBase.ts
@@ -1,0 +1,6 @@
+export const padBase64Str = (str: string) => {
+  while (str.length % 4 !== 0) {
+    str += '=';
+  }
+  return str;
+};
